rcll: do not sync keep-mps-side-free facts
The rules that maintain these facts are solely depended on facts
available locally on each agent. Thus the fact does not rely on syncing
to be consistent across the agents. However, because the fact is
in some scenarios short-lived (less than 0.1s) it can cause syncing
issues leading to inconsistent world states. Thus it is better to
remove the syncing and just rely on the local computation.