rcll: track step after next step for deadlock avoidance in with promises
With promises not only the next step in the production chain of a
product is relevant, but also the step after, since the relevant changes
might not be applied yet leading to the tracking lagging behind.
Therefore the step after the next step is tracked now too to allow for
better lookahead.