rcll: keep rs-side free looking one step ahead
With promises a new kind of deadlock can occur at the ring stations
which can be prevented by looking one step ahead. This commit adds a
rule to assert a corresponding keep-mps-side-free fact if such a
situation occurs. Concretely this can happen if a workpiece is at the
input of an RS and after the current ring is mounted it has another
ring to be mounted at the same machine.