rcll: prevent duplicate fill cap goals on multiple agents
Promises make it possible for multiple agents to try to fill the same
cap station, breaking the machine. Since we do not share information
about which goals are executed on which agent, we use shared facts
to deduce that such situations occur and prevent them by modifying
the precondition for FILL-CAP goal formulation.