rcll-central: merge open and active discard+payment requests
Previously only an active + an open request for discard and payment with
base could be merged into one payment with cap carrier. However, this
limitation was not necessary and reduced the potential impact of the
merge. Therefore the rule is changed to only work on active requests,
which means that open requests will be converted into their base case
first and are then automatically active. Then this rule can merge them
with other active requests.