goal-class: add new goal that picks up workpieces from machine outputs
If the output needs to be free for workpieces to be put on the input
of a machine, then deadlocks can occur, e.g. by two RS outputs being
full but the workpieces need to go into the respective different RS.

To solve such situations, a workpiece can simply be picked up. This
new goal attempts this.