promises: manage lock resources through promises to prevent deadlocks
Lock actions are used to guarantee mutually exclusive access to resources.
Because of their nature, promised goals tend to access the same resources
as their promising goals. If a promised goal wants to access such a
resource at the beginning of its plan, while the promising goal accesses
the resource at the end of its plan, the promised goal might acquire the
lock first. However, since the promising goal has not finished execution,
this might cause the promised changes to the world to not have been
applied yet. Thus the promised goal might stuck in the locked state,
blocking both the promising goal and the promised goal until the promised
goal eventually times out.