gazsim: overwrite clips-specs configurations based on instance config
Switching between agent modes (central, decentral) not only requires
different meta plugins to be loaded but also configuring the right
clips-spec. While the former can be done through fawkes and gazsim
parameters (and thus be automatically changed without changing the
contents of files), the clips-specs are configured in
clips-executive.yaml and thus require a file change. However, there
is also default configuration for the agents for each of the modes
of operation. Thus one way to avoid the requirement of changing a file
is to set the config value in the default config files, which are
loaded according to the mode. However, the config value must also be
removed from clips-executive.yaml as it is not correctly overwritten
otherwise.