simtests: configure new simtests for the build tests
The buildtests for central and decentral agent will be reconfigured
to require the three simtests configured here:
ENTER-FIELD serves as a basic unit tests to check agent startup.
PRODUCE-C1 serves as a unit test for the mainline production.
FULL serves as a test to measure agent capability against a fixed and
reasonable minimum point threshold.

Fixed games and an adapted testing script will be added in later PRs.