rcll-central: add prototype implementation of robot-bound order asgnmt
Currently we restricted starting the production of orders through a
series of filters and then dynamically assign both production and
support goals necessary for that order dynamically based on priorities
and executability. An alternative strategy to this would statically
assign robots to one of two roles: support or producer. A producer
takes care of all the mainline production goals of just one statically
assigned order. Once the order is completed, it can pick a new order
from the available ones (i.e. according to a simple heuristic based
on delivery-end time and order complexity, to maximize points). The
support robots would dynamically switch between goals of different
orders based on priority and executability to fulfill the support goal
requirements of the active orders (e.g. buffering caps or making
payments).