rcll-central: fix order workload calculation
The calculation for order-caused workload for a machine was broken
in the sense that it correctly only did not consider already fulfilled
actions but incorrectly considered all non-fulfilled orders, even
those not started yet. This led to a steady workload increase to the
point where no order could be pursued anymore.