rcll-central: allow wp-get and wp-check when MPS broken
Previously, we failed a goal in wp-get if the machine breaks because
we could not be sure if we actually grabbed it. Now with wp-check
we have that ability, and therefore wont fail the goal in this case
anymore. If a judge says we need to not use that WP anymore, we need
to take the robot into maintenance anyhow, which resets the WM.