rcll-central: add cool down timeout after goal fails repeatedly
In some cases a robot continuously fails with a goal due to no
reason intrinsic to the goal (e.g. hardware failure). After repeated
failure, the goal might be blocked for a while from executability
to either allow other robots to take on the goal (if possible) or
to give the robot a completely new starting point towards the goal
since it would start another goal in between.