rcll-central: make wait for points configurable
Waiting for points allows us to wait until we sense points for certain
actions (like slide puts) as some form of (incomplete) environmental
feedback on action success. However, they are not supported by the
simulation. Therefore make them configurable and disable them in
the host specific config of the simulation.