rcll-central: begin reworking requests
The idea is to not use requess in goal formulation.
Rather have all possible goals be formulated (e.g., options to pay with
a cap-carrier or to discard it, options to buffer caps, options to pay
on each ring station.
Requests should merely state the need for certain operations caused by a
respective order.
That way they only affect prioritization of existing goals.
If requests are messed up, robots still have the possibility to do
payments and such, hopefully preventing deadlocks.