rcll-central: add mode for new rcll-get-started + protobuf sim
The idea is to have a setup that can directly start the agent without
much other stuff (no mongodb, no fawkes instances for other robots) that
does not even need gazsim.bash.

It also adapts the agent to interface with the updated protobuf sim
which has separate peers for different robots and periodically sends
responses on running and old tasks.